+++ src/main/java/com/salesianostriana/dam/TrailQuest_Api/dto/route/RouteResponseDTO.java
@@ -1,8 +1,7 @@
-package com.salesianostriana.dam.TrailQuest_Api.dto;
+package com.salesianostriana.dam.TrailQuest_Api.dto.route;
 
 import com.salesianostriana.dam.TrailQuest_Api.model.PosiblesRegiones;
 import com.salesianostriana.dam.TrailQuest_Api.model.Route;
-import jakarta.validation.constraints.NotNull;
 
 import java.util.UUID;
